README.md: Fix the quick start example
- Quote the heredoc delimiter so that variable expansion does not occur.

Prior to this change, in the example that shows how to create
credential.txt, the here-document delimiter was unquoted which would
subject all the lines to unintended variable expansion. For example, if
the user's content contained something like abc$defg then $defg would be
wrongly treated as a variable instead of literally.

Ref: https://stackoverflow.com/a/22698106